(refactor) combined the urls into one object
