+++ src/applications/check-in/day-of/routes.jsx
@@ -17,7 +17,7 @@ import withFeatureFlip from './containers/withFeatureFlip';
 import withSession from './containers/withSession';
 import withToken from './containers/withToken';
 import withForm from '../containers/withForm';
-import { URLS } from '../utils/navigation/day-of';
+import { URLS } from '../utils/navigation';
 
 const routes = [
   {
